fix: local PySpark bootstrap on Windows + EIA dtype
Makes the pipeline runnable end-to-end on Windows from a cold venv:

  - scripts/env.sh: export HADOOP_HOME (→ vendor/hadoop, winutils-3.3.5),
    PYSPARK_PYTHON and PYSPARK_DRIVER_PYTHON so the JVM can fork
    Python workers inside the venv.
  - .gitignore: skip vendor/hadoop (Windows-only, downloaded by script).
  - src/ingestion/eia_client.py: cast load_mw to float64 — EIA returns
    plain ints which clash with the Bronze DoubleType schema.

Verified end-to-end on CAISO (CAL) for 2024-04-01..2026-04-22:
  - Bronze:   18,025 rows
  - Silver:   18,022 rows (IQR 8,483..52,163 MW — matches CA peaks)
  - Gold:     17,854 rows × 32 features
  - LightGBM: MAPE 1.81% / RMSE 700 MW / MAE 533 MW
  - Isolation Forest: 179 anomalies (1.00%)
  - Registered energy-demand-forecaster@staging v1
  - Registered energy-anomaly-detector@staging v1

+++ src/ingestion/eia_client.py
@@ -160,7 +160,9 @@ class EiaClient:
 
         df = pd.DataFrame(data)
         df["timestamp_utc"] = pd.to_datetime(df["period"], utc=True)
-        df["load_mw"] = pd.to_numeric(df["value"], errors="coerce")
+        # Force float64 — EIA sometimes returns integers, but the Bronze
+        # Delta schema pins DoubleType.
+        df["load_mw"] = pd.to_numeric(df["value"], errors="coerce").astype("float64")
         return df[["timestamp_utc", "load_mw"]].dropna()
 
     def close(self) -> None:
